set up hosting on heroku at https://our-places-challenge.herokuapp.com/

+++ backend/app.js
@@ -29,30 +29,37 @@ app.use((req, res, next) => {
 app.use("/api/places", placesRoutes);
 app.use("/api/users", usersRoutes);
 
-app.use((req, res, next) => {
-  const error = new HttpError("Could not find this route.", 404);
-  throw error;
-});
+if (process.env.NODE_ENV === 'production') {
+    app.use(express.static('../frontend/build'));
+    app.get('*', (req, res) => {
+      res.sendFile(path.resolve(__dirname, '..', 'frontend', 'build', 'index.html'));
+    })
+}
+else {
+  app.use((req, res, next) => {
+    const error = new HttpError("Could not find this route.", 404);
+    throw error;
+  });
 
-app.use((error, req, res, next) => {
-  if (req.file) {
-    fs.unlink(req.file.path, (err) => {
-      console.log(err);
-    });
-  }
-  if (res.headerSent) {
-    return next(error);
-  }
-  res.status(error.code || 500);
-  res.json({ message: error.message || "An unknown error occurred!" });
-});
+  app.use((error, req, res, next) => {
+    if (req.file) {
+      fs.unlink(req.file.path, (err) => {
+        console.log(err);
+      });
+    }
+    if (res.headerSent) {
+      return next(error);
+    }
+    res.status(error.code || 500);
+    res.json({ message: error.message || "An unknown error occurred!" });
+  });
+}
 
 mongoose
-  .connect(
-    `mongodb+srv://<username>:<password>@c<mongodb_service>/<dataabaseName>?retryWrites=true&w=majority`
+  .connect(process.env.MONGODB_URI
   )
   .then(() => {
-    app.listen(5000);
+    app.listen(process.env.PORT || 5000);
   })
   .catch((err) => {
     console.log(err);
